Refactor StockInfoDto 패키지 이전 (infrastructure -> dto)
StockInfoDto 가 전 레이어에서 공통적으로 사용되어 dto 자체를 다른 레이어들과 같은 depth 에 위치해도 좋을 것 같아 이전합니다.
