+++ ktb-be-dev-assignment/src/main/java/org/ktb/ktbbedevassignment/dto/StockInfoDto.java
@@ -1,4 +1,4 @@
-package org.ktb.ktbbedevassignment.infrastructure;
+package org.ktb.ktbbedevassignment.dto;
 
 public record StockInfoDto(String companyName, String tradeDate, float closingPrice) {
 }
