+++ ktb-be-dev-assignment/src/main/java/org/ktb/ktbbedevassignment/infrastructure/StockRepository.java
@@ -1,5 +1,6 @@
 package org.ktb.ktbbedevassignment.infrastructure;
 
+import org.ktb.ktbbedevassignment.dto.StockInfoDto;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
