+++ ktb-be-dev-assignment/src/test/java/org/ktb/ktbbedevassignment/application/StockServiceTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.ktb.ktbbedevassignment.exception.CompanyNotFoundException;
 import org.ktb.ktbbedevassignment.infrastructure.CompanyRepository;
-import org.ktb.ktbbedevassignment.infrastructure.StockInfoDto;
+import org.ktb.ktbbedevassignment.dto.StockInfoDto;
 import org.ktb.ktbbedevassignment.infrastructure.StockRepository;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
