+++ ktb-be-dev-assignment/src/test/java/org/ktb/ktbbedevassignment/infrastructure/StockRepositoryTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
+import org.ktb.ktbbedevassignment.dto.StockInfoDto;
 import org.ktb.ktbbedevassignment.util.TestDataHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
